UI package refactor
