+++ internal/ui/controls.go
@@ -1,4 +1,4 @@
-package main
+package ui
 
 import tea "github.com/charmbracelet/bubbletea"
