+++ internal/ui/plex_album_browse.go
@@ -1,4 +1,4 @@
-package main
+package ui
 
 import (
 	"fmt"
